+++ service/app/controllers/courseenrollment/validator/CourseEnrollmentRequestValidator.java
@@ -21,11 +21,11 @@ public class CourseEnrollmentRequestValidator extends BaseRequestValidator {
     validateParam(
         (String) courseRequestDto.getRequest().get(JsonKey.COURSE_ID),
         ResponseCode.mandatoryParamsMissing,
-        JsonKey.COURSE_ID+"/"+JsonKey.ENROLLABLE_ITEM_ID+"/"+JsonKey.COLLECTION_ID);
+        JsonKey.COURSE_ID+"/"+JsonKey.COLLECTION_ID);
     validateParam(
         (String) courseRequestDto.getRequest().get(JsonKey.BATCH_ID),
         ResponseCode.mandatoryParamsMissing,
-        JsonKey.BATCH_ID+"/"+JsonKey.FIXED_BATCH_ID);
+        JsonKey.BATCH_ID);
     validateParam(
         (String) courseRequestDto.getRequest().get(JsonKey.USER_ID),
         ResponseCode.mandatoryParamsMissing,
